+++ babel-compiler-hot/package.js
@@ -1,14 +1,14 @@
 Package.describe({
   name: "gadicc:babel-compiler-hot",
-  // version: '6.5.0-beta.12'    // core version, KEEP UPDATED
-  version: '0.0.6-beta.12',
+  // version: '6.5.1-rc.1'    // core version, KEEP UPDATED
+  version: '0.0.7-rc.1',
   summary: 'React hotloading, used by gadicc:ecmascript-hot.',
   git: 'https://github.com/gadicc/meteor-react-hotloader',
   documentation: 'README.md'
 });
 
 Npm.depends({
-  'meteor-babel': '0.8.0',
+  'meteor-babel': '0.8.2',
   'mongodb': '2.1.4',
   'babel-plugin-react-transform': '2.0.0'
 });