+++ modules-runtime-hot/package.js
@@ -1,7 +1,7 @@
 Package.describe({
   name: "gadicc:modules-runtime-hot",
-  // version: "0.5.0-beta.11",    // core version, KEEP UPDATED
-  version: "0.0.3-beta.12",
+  // version: "0.6.1-rc.1",    // core version, KEEP UPDATED
+  version: "0.0.4-rc.1",
   summary: 'Used by gadicc:ecmascript-hot',
   git: "https://github.com/benjamn/install",
   documentation: "README.md"
@@ -9,7 +9,7 @@ Package.describe({
 
 /*
 Npm.depends({
-  install: "0.5.4"    // core version, KEEP UPDATED
+  install: "0.6.0"    // core version, KEEP UPDATED
 });
 */
 
@@ -33,7 +33,7 @@ Package.onUse(function(api) {
   api.export("meteorInstall");
 
   // hot
-  api.versionsFrom('1.3-beta.12');
+  api.versionsFrom('1.3-rc.1');
   api.use('webapp', 'server');
   api.addFiles('modules-runtime-proxy.js', 'server');
 });