+++ ecmascript-hot/package.js
@@ -1,7 +1,7 @@
 Package.describe({
   name: 'gadicc:ecmascript-hot',
   // version: '0.4.1-rc.2'    // core version, KEEP UPDATED
-  version: '0.0.8-rc.2',
+  version: '0.0.9-rc.2',
   summary: 'Replacement ecmascript package providing react hotloading',
   git: 'https://github.com/gadicc/meteor-react-hotloader',
   documentation: '../README.md'
@@ -15,20 +15,19 @@ Package.registerBuildPlugin({
 
 Package.onUse(function (api) {
   api.versionsFrom('1.3-rc.2');
-  
   api.use('isobuild:compiler-plugin@1.0.0');
-  api.use('gadicc:babel-compiler-hot@0.0.8-rc.2');
+  api.use('gadicc:babel-compiler-hot@0.0.11-rc.2');
 
   api.imply('modules');
   api.imply('ecmascript-runtime');
   api.imply('babel-runtime');
   api.imply('promise');
 
-  api.use('gadicc:hot@0.0.7');
+  api.use('gadicc:hot@0.0.8');
   api.imply('gadicc:hot');
 
   api.addFiles("ecmascript.js", "server");
-  api.export("ECMAScript", 'server');
+  api.export("ECMAScript", "server");
 });
 
 Package.onTest(function (api) {