+++ History.md
@@ -1,5 +1,7 @@
 # vNEXT
 
+* Release for Meteor `1.3-rc.1`
+
 # v0.0.7-beta.12 (2016-03-09)
 
 * Accept function components that contain code (before we just accepted functions